chore: factor out electron main menu
